+++ tests/testthat/test_basic_ops.R
@@ -27,3 +27,24 @@ test_that("valid DataBackend (as_sqlite_backend)", {
   expect_iris_backend(b, n_missing = 30L)
   DBI::dbDisconnect(private(b)$.data$src$con)
 })
+
+test_that("strings_as_factors", {
+  data = iris
+  tbl = as_sqlite_tbl(data)
+
+  b = DataBackendDplyr$new(data = tbl, "row_id", strings_as_factors = FALSE)
+  expect_character(b$head()$Species, any.missing = FALSE)
+  expect_character(b$data(b$rownames[1], "Species")$Species, any.missing = FALSE)
+
+  b = DataBackendDplyr$new(data = tbl, "row_id", strings_as_factors = TRUE)
+  expect_factor(b$head()$Species, any.missing = FALSE)
+  expect_factor(b$data(b$rownames[1], "Species")$Species, any.missing = FALSE)
+
+  b = DataBackendDplyr$new(data = tbl, "row_id", strings_as_factors = "Species")
+  expect_factor(b$head()$Species, any.missing = FALSE)
+  expect_factor(b$data(b$rownames[1], "Species")$Species, any.missing = FALSE)
+
+  expect_error(DataBackendDplyr$new(data = tbl, "row_id", strings_as_factors = "Sepal.Length"))
+
+  DBI::dbDisconnect(private(b)$.data$src$con)
+})