+++ NEWS.md
@@ -1,6 +1,7 @@
 # mlr3db 0.1.2
 
-* Fixed `$distinct()` to not return missing values.
+* Fixed `$distinct()` to not return missing values per default.
+* Added `na_rm` argument to `$distinct()`.
 
 # mlr3db 0.1.1